pyproject.toml
[tool.poetry]
name = "hazid-project3"
version = "0.1.0"
description = ""
authors = ["Daulet <dauletkhabidullin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-multiselectfield = "^0.1.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
